adjusted more front end navigation and formated navigation code
