+++ src/Navigation/FrontEndNavigation/AngularNavigation/BestPractices.ts
@@ -3,11 +3,11 @@ import { Subpage } from "@/Navigation/NavigationTypes";
 const AngularBestPracticesNavigation: Subpage = {
   name: "Angular Best Practices",
   subpages: [
-        {
-          name: "Best Practices", // Combined and shortened
-          path: "/angular/best-practices",
-        },
-      ],
+    {
+      name: "Best Practices", // Combined and shortened
+      path: "/angular/best-practices",
+    },
+  ],
 };
 
 export default AngularBestPracticesNavigation;
