+++ src/Navigation/FrontEndNavigation/AngularNavigation/CoreConcepts.ts
@@ -3,23 +3,23 @@ import { Subpage } from "@/Navigation/NavigationTypes";
 const AngularCoreConceptsNavigation: Subpage = {
   name: " Angular Core Concepts",
   subpages: [
-        {
-          name: "Components",
-          path: "/angular/coreconcepts/components",
-        },
-        {
-          name: "Modules",
-          path: "/angular/coreconcepts/modules",
-        },
-        {
-          name: "Services & DI",
-          path: "/angular/coreconcepts/services",
-        },
-        {
-          name: "Routing",
-          path: "/angular/coreconcepts/routing",
-        },
-      ],
+    {
+      name: "Components",
+      path: "/angular/coreconcepts/components",
+    },
+    {
+      name: "Modules",
+      path: "/angular/coreconcepts/modules",
+    },
+    {
+      name: "Services & DI",
+      path: "/angular/coreconcepts/services",
+    },
+    {
+      name: "Routing",
+      path: "/angular/coreconcepts/routing",
+    },
+  ],
 };
 
 export default AngularCoreConceptsNavigation;
