+++ src/Navigation/FrontEndNavigation/HTMLNavigation/HTMLBestPractices.ts
@@ -1,0 +1,17 @@
+import { Subpage } from "@/Navigation/NavigationTypes";
+
+const HTMLNavigation: Subpage = {
+  name: "HTML Best Practices",
+  subpages: [
+        {
+          name: "Code Standards",
+          path: "/html/best-practices/standards", // Clean and maintainable HTML
+        },
+        {
+          name: "Performance Optimization",
+          path: "/html/best-practices/performance", // Techniques to enhance performance
+        },
+      ],
+};
+
+export default HTMLNavigation;
