+++ src/Navigation/FrontEndNavigation/HTMLNavigation/HTMLIntermediate.ts
@@ -1,0 +1,17 @@
+import { Subpage } from "@/Navigation/NavigationTypes";
+
+const HTMLNavigation: Subpage = {
+  name: "HTML Intermediate",
+  subpages: [
+    {
+      name: "Media",
+      path: "/html/intermediate/media", // Embedding images, videos, and audio
+    },
+    {
+      name: "Tables",
+      path: "/html/intermediate/tables", // Creating and styling tables
+    },
+  ],
+};
+
+export default HTMLNavigation;
