+++ src/Navigation/FrontEndNavigation/JavaScriptPages/Basics.ts
@@ -1,0 +1,14 @@
+import { Subpage } from "@/Navigation/NavigationTypes";
+
+const JavaScriptBasicsNavigation: Subpage = {
+  name: "JavaScript Basics",
+  subpages: [
+    { name: "Overview", path: "/javascript/basics/overview" }, // Introduction and history
+    { name: "Syntax", path: "/javascript/basics/syntax" }, // Basic structure
+    { name: "Data Types", path: "/javascript/basics/datatypes" }, // Strings, numbers, objects, etc.
+    { name: "Variables", path: "/javascript/basics/variables" }, // let, const, var
+    { name: "Functions", path: "/javascript/basics/functions" }, // Function declarations and expressions
+  ],
+};
+
+export default JavaScriptBasicsNavigation;
